Add input flag aliases and improve docs
Adds input flags -J for json and -T for toml.
These do not clash with jq's cli flags.

A potentially scary thing about this approach is that we could not do this for any new format.
E.g. if we added RON support, we would not be able to add -R shorthand for ron input.

Thinking that these (less primary, potential future formats) will just have to deal with the explicit --input=ron style.

We could of course do the coreutils binary style thing, but the short xq style binary names are already very much taken.
Soo, that would leave us with lqt lqy lqj etc for input, or lqyy lqjt, or lq y, lq j, lq jt. Not nice.

In the end, going for this micro-optimisation because i write tools for myself that i want to be nice to use now, rather than idealistic for a future that probably never comes.

Signed-off-by: clux <sszynrae@gmail.com>